+++ src/control_plane/shared_infrastructure/rabbitmq/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) use rabbitmq_definitions::RabbitMqDefinitions;
+pub(crate) use rabbitmq_password_hash::RabbitMqPasswordHash;
+pub(crate) use rabbitmq_plan_error::RabbitMqPlanError;
+pub(crate) use rabbitmq_project_definition::RabbitMqProjectDefinition;
+
+mod rabbitmq_definitions;
+mod rabbitmq_password_hash;
+mod rabbitmq_plan_error;
+mod rabbitmq_project_definition;